Let conf.parse() open the file from module entry point
Instead of opening the file in conf._main(), we let conf.parse() handle this.
This is necessary to provide access to the actual file path when using the
module entry point in preparation for adding support to parse include
directives in the next commits (where knowledge of the initial file path is
needed).