Add support for include directives in configuration files
Method parse() of Configuration class is changed to return a list of include
directives found in processed file (previously, these were simply considered as
entries, now they are not). This list is then processed in the main parse()
function in 'conf' module by possibly recursing through include directives
found in included configuration file while they are processed. The kind of
include directives is modelled with an IncludeType enum.

In tests, we check for error cases and also use the 'data/postgres.conf' file
as a functional test case.