Handle commented configuration values
We add a 'commented' field in Entry objects to store whether a parsed line
is a commented parameter or an actual value. The logic is handled in
Configuration.parse(): we distinguish real comments from commented parameters
depending on whether we can actually parse the uncommented line.

Handling of includes directive needs adjustments because commented lines
containing such a value should be ignored (i.e. included files or directories
not loaded): change in data/postgres.conf covers this case.

When editing a Configuration object, either through __setitem__ or through the
edit() context manager, we uncommented edited values.