Rewrite parse logic of configuration include directives
The previous logic was wrong because included files were parsed after the
primary file was completely parsed instead of being parsed relative to their
location in that file.

We rewrite conf.parse() by using a generator pattern that yield parsing of an
include directive when it is encountered to a parse_include() sub-generator.

Wrong test expectations are now fixed.