Add an async interface for PGCtl
We add the AsyncPGCtl class and factor out code common to sync and async
interfaces into an AbstractPGCtl class. Since construction of the PGCtl object
involves some I/O (in order to set the 'version' field), we handle this in a
get() class method in the async case.

In tests, we have to duplicate cases for sync and async interfaces, although
not all as we now rely on the factorization in AbstractPGCtl to check _cmd()
methods instead of actual I/O methods. Functional tests are moved to a new
module, in order to avoid confusing fixtures in particular.

We no longer run tests during Debian package build, as these fail with
pytest-asyncio plugin which is too old in CI images.