Fix connection service serialization wrongly including 'name' parameter
The Service class overwrite __setattr__ to invoke __setitem__ so the statement
'self.name = name' in Service.__init__() will actually add the 'name' key to
the Service (which is a dict). Then, as serialization invokes .items() to get
service parameters, the 'name' field is written whereas is should not.

We invoke __setattr__() from super() in order to avoid that and also declare
the 'name' attribute in the class in order to keep type checking working.