Only append uncommented entries when editing pgconf
Follow-up commit 7604544e1c170dc8855facd8532543a9630f24bc. With the test case
as is now, and the previous code, we'd see a '# bar = off' entry in updated
file, which does not make sense as the (uncommented) value from included file
would apply. Accordingly, we only append the new entry if uncommented.