Fix broken upload to Codecov
Codecov now requires a TOKEN to upload results.
We update the version of the github action and use the token added in the
project settings.

We write the coverage report as XML in a file for codecov upload to be happy.
With the default .coverage file, codecov reports a "no coverage file found"
error.